n8n/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
boto3
